fix: Remove try catch of repositories and use cases return responses

+++ backend/src/use-cases/user/delete-user-by-id.use-case.ts
@@ -1,16 +1,22 @@
+import { DeleteUserResponse } from 'src/dtos';
 import { UserRepository } from 'src/repositories';
 import { NotFoundException } from '@nestjs/common';
 
 export class DeleteUserByIdUseCase {
   constructor(private readonly userRepository: UserRepository) {}
 
-  async execute(id: number): Promise<void> {
+  async execute(id: number) {
+    const user = await this.userRepository.findById(id);
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     const hasDeleted = await this.userRepository.delete(id);
 
     if (!hasDeleted) {
       throw new NotFoundException('User not found');
     }
 
-    return;
+    return DeleteUserResponse.fromEntity(user);
   }
 }
